pyproject.toml
[project]
name = "garage"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AndreRichartz", email = "AndreRichartz"},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
